requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
pytest>=7.4.0
colorama>=0.4.6
tqdm>=4.65.0
python-dateutil>=2.8.2
urllib3>=2.0.0
aiohttp>=3.8.0